// src/types/tapWaterBusiness.ts
export type ChangeMetrics = Record<string, { change?: number; percentChange?: number }>;
export type Rolling12MonthSums = {
  distributedWaterVolumeM3?: number;
  billedWaterVolumeM3?: number;
  taiwanWaterSupportVolumeM3?: number;
  waterUseExcludingTaiwanWaterSupportM3?: number;
  monthlyRevenueThousandNtd?: number;
  monthlyExpenseThousandNtd?: number;
  monthlySurplusThousandNtd?: number;
};

export type TapWaterBusinessKeyMetricRecord = {
  id: string;
  module: 'tap_water_business_key_metrics';
  periodRaw: string;
  rocYear: number;
  year: number;
  month: number;
  monthKey: string;
  periodDate: string;
  quarter: string;
  yearMonthLabel: string;
  isLatestMonth: boolean;
  distributedWaterVolumeM3?: number;
  directDrinkingFountainCount?: number;
  onlineWaterQualityMonitoringStationCount?: number;
  averageWaterPressureKgCm2?: number;
  transmissionDistributionPipelineLengthKm?: number;
  billedWaterVolumeM3?: number;
  taiwanWaterSupportVolumeM3?: number;
  waterUseExcludingTaiwanWaterSupportM3?: number;
  supplyAreaTotalPopulation?: number;
  supplyAreaServedPopulation?: number;
  supplyCoverageRatePercent?: number;
  userCount?: number;
  perCapitaDailyWaterUseLiter?: number;
  employeeCount?: number;
  staffCount?: number;
  workerCount?: number;
  usersPerEmployee?: number;
  monthlyRevenueThousandNtd?: number;
  salesRevenueThousandNtd?: number;
  serviceRevenueThousandNtd?: number;
  otherOperatingRevenueThousandNtd?: number;
  nonOperatingRevenueThousandNtd?: number;
  monthlyExpenseThousandNtd?: number;
  salesCostThousandNtd?: number;
  serviceCostThousandNtd?: number;
  otherOperatingCostThousandNtd?: number;
  nonOperatingExpenseThousandNtd?: number;
  monthlySurplusThousandNtd?: number;
  assetsThousandNtd?: number;
  liabilitiesThousandNtd?: number;
  equityThousandNtd?: number;
  distributedWaterVolumeMillionM3?: number;
  billedWaterVolumeMillionM3?: number;
  taiwanWaterSupportVolumeMillionM3?: number;
  waterUseExcludingSupportMillionM3?: number;
  supportShareOfDistributedWaterPercent?: number;
  billedShareOfDistributedWaterPercent?: number;
  nonRevenueOrUnbilledWaterApproxM3?: number;
  nonRevenueOrUnbilledWaterApproxPercent?: number;
  populationServedSharePercent?: number;
  usersPerThousandPopulation?: number;
  monthlyRevenueMillionNtd?: number;
  monthlyExpenseMillionNtd?: number;
  monthlySurplusMillionNtd?: number;
  operatingRevenueThousandNtd?: number;
  operatingExpenseThousandNtd?: number;
  operatingMarginApproxPercent?: number;
  debtRatioPercent?: number;
  equityRatioPercent?: number;
  monthOverMonthChanges?: ChangeMetrics;
  yearOverYearChanges?: ChangeMetrics;
  rolling12MonthSums?: Rolling12MonthSums;
  sourceRecordHash: string;
  source: string;
  sourceAgency: string;
};

export type TapWaterBusinessKeyMetricSummary = {
  totalRecords: number;
  minMonthKey?: string;
  maxMonthKey?: string;
  latestMonthKey?: string;
  latest?: Partial<TapWaterBusinessKeyMetricRecord>;
  byYear: Array<{
    year: number;
    monthCount: number;
    distributedWaterVolumeM3Sum?: number;
    billedWaterVolumeM3Sum?: number;
    taiwanWaterSupportVolumeM3Sum?: number;
    waterUseExcludingTaiwanWaterSupportM3Sum?: number;
    monthlyRevenueThousandNtdSum?: number;
    monthlyExpenseThousandNtdSum?: number;
    monthlySurplusThousandNtdSum?: number;
    averagePerCapitaDailyWaterUseLiter?: number;
    averageSupplyCoverageRatePercent?: number;
  }>;
  rolling12MonthLatest?: { monthKey: string } & Rolling12MonthSums;
  metricExtremes: Record<string, { monthKey: string; value: number } | undefined>;
  dataQuality: {
    invalidPeriodCount: number;
    missingMetricValueCount: number;
    invalidNumericValueCount: number;
    duplicateMonthKeyCount: number;
  };
};